Set dbname and username when MyProcPort is not initialized when setup connection.
InitPostgres() in GPDB needs to recover DTM, and will setup connection
from QD to QE by calling cdbconn_doConnectStart(), cdbconn_doConnectStart()
uses MyProcPort to get username and dbname currently. But for background
worker, MyProcPort is not set. Set MyProcPort before InitPostgres() is not
appropriate since catalog is not initialized. We could use MyDatabaseId and
AuthenticatedUserId instead, which are initialized in InitPostgres() for
background worker.