Don't bind outgoing TCP connections to a particular IP address.
In the TCP interconnect, we used to bind outgoing TCP connections to the
same source IP address that the libpq connection came from. That can lead
to running out of ephemeral TCP ports. I was seeing errors, when running
the regression tests with the TCP interconnect:

ERROR:  interconnect error setting up outgoing connection
DETAIL:  Could not bind to local addr 2.0.0.0: Address already in use

This was easily reproducible, by running the parallel group of tests that
includes the qp_misc_jiras test. Apparantly that parallel group opens
especially many connections.

When a socket is bound to a particular IP address with bind(), it is also
allocated an ephemeral TCP port. On Linux, the range of ports available
can be seen in /proc/sys/net/ipv4/ip_local_port_range. It defaults to
32768-60999, but even if you increase the range, it's always quite
limited. bind() reserves the whole TCP port, even though multiple outgoing
connections could share the same source port, as long as their destination
IP address or port is different, because bind() doesn't know whether
you're going to use the port to listen for incoming connections, or for
establishing an outgoing connection. Listening for incoming connections
needs to reserve the port.

Linux kernel 4.2 introduced a new socket option, IP_BIND_ADDRESS_NO_PORT,
that we could use to give bind() a hint that we're using the socket for
an outgoing connection, so there's no need to reserve the whole port. But
actually, I don't think we should be calling bind() on outgoing
connections in the first place. I don't think the logic, to use the
incoming libpq connection's IP address as the source IP address of outgoing
interconnect connections, makes sense. The comment says that it is for
fault tolerance, but I don't buy that argument. If a network adapter is not
working, it should be disabled in the OS configuration so that it is not
used. It is not the application's job to make routing decisions.

Forcing the same source IP address seems outright wrong in some scenarios.
Imagine that the QD has two network adapters: one for connecting to the
outside world, and another for the internal network where the QEs are. In
that scenario, the interconnect connections between the QD and the QEs
should definitely *not* be established through the same network adapter as
the user's libpq connection.

Better to just remove the code to bind to a particular source IP address,
and let the OS do its job of routing TCP connections. (AFAICS, the UDP
interconnect never tried to force a particular source IP address when
sending.)

Reviewed-by: Paul Guo <pguo@pivotal.io>
Discussion: https://groups.google.com/a/greenplum.org/d/msg/gpdb-dev/ITkZdACpcVQ/H_74phbMFgAJ

Conflicts:
	gpAux/extensions/googletest
	src/backend/cdb/motion/ic_tcp.c