Dispatch GUCs to QE along with query.
In past, GPDB dispatch GUCs mainly in two ways:
1. Set related command to set GUC value and dispatch
"SET COMMAND" to QEs.
2. Pass GUCs with flag GUC_GPDB_ADDOPT to QEs when creategang.

It has two problems:
1. Set command is not 2PC and may lead to inconsistent GUC state.
Issue-6610 describe the details of this problem.
2. When set GUCs without GUC_GPDB_ADDOPT using SET command, it will
apply to all the existing QEs, but when cached QE exits with timeout,
the changed GUC value will not appear on the new QE process.

Make set command 2PC is too heavy, which is discussed in PR-6596.
To fix it without 2PC, this PR follow the way to dispatch GUCs to
QEs when GUC state is not consistent. Flag guc_need_sync is used
to determine whether GUCs are not in consistent state between
QD and QE. When GUC value changed or transaction abort, this flag
will be turned on.
Instead of passing all the GUCs with flag GUC_GPDB_ADDOPT, we track
the changed GUC list and only pass this subset.

Idea from Heikki Linnakangas <hlinnakangas@pivotal.io>
Idea from Pengzhou Tang <ptang@pivotal.io>