Fix bug: initplan may clean up gangs allocated to parent plan
When initplan finished, it will call the function freegangsforportal,
this function recycles or destroys the allocated gangs (including
those gangs allocated to parent plan). The might lead to invalid
memory reference when later dispatch parent plan to its gangs.

In this commit, we add a field `dispatched` in the Gang struct,
if this field is false, the gang will not be destroyed.

Co-authored-by: Hubert Zhang <hzhang@pivotal.io>