Eliminating alien nodes before execution on master.
GPDB support to eliminate alien nodes for QEs, but master still
needs to run ExecInitNode for all the plan node. This will
introduce unneccessary CPU cost on master.

For non explain and explain analyze query, master only needs to
ExecInitNode for all node on top of the first gather motion along
with initplan which belongs to master.