Improve the performance of large number of small table redistribution
To modify the way of recording redistribution information, previously
using UPDATE, there was a concurrent lock conflict problem. Although
GDD was introduced in version 6, there is still a concurrent UPDATE
lock conflict problem in the environment where GDD is not turned on.
In this modification, INSERT is used to recording the redistribution
process and avoid the lock conflict of UPDATE