Fix flaky test 'gangsize'
In the test case, there's query like 'insert into t select i from
generate_series(1,10) i', the slice of 'generate_series' has the locus of
general, so it might be executed in any segment according to the session id and
that makes the test flaky. To make it deterministic, we change generate_series
to a regular table and filter the data with gp_segment_id. This commit also
removes the alternative expect files.

Co-authored-by: Gang Xiong <gangx@vmware.com>