ic-proxy: fix address classifying
The ic-proxy addresses specified with the GUC
gp_interconnect_proxy_addresses must be classified as removed, added,
updated, and unchanged.  This is done by comparing the old address list
and the new list.  However an bug was that only the dns-resolved
addresses were used as the old ones, so if the address reloading logic
is re-triggered before all the addresses are resolved, then the
unresolved addresses will be marked as removed unexpectedly.

Fixed by combining both the dns resolved and unresolved addresses as the
old ones.