Fix the flaky test of autovacuum
current test case assume there is only one autovacuum worker for a specific
database at the same time. But when the worker is blocked, i.e. by the fault
inject, the autovacuum laucher will start a new worker to vacuum this database.
Discussed in gpdb-dev:
https://groups.google.com/a/greenplum.org/g/gpdb-dev/c/8QNIiwtjh2U/m/8TLN2UKjBAAJ
It's better to the autovacuum behavior the same as upstream. Hence, we update
the test case by waiting all the three(decide by GUC autovacuum_max_workers)
workers to be finished.